requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.2.0
Django==5.1rc1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Faker==30.1.0
fpdf==1.7.2
gunicorn==23.0.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
reportlab==4.2.5
segno==1.6.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.7.0
